Relative imports and restructuring
